# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gorilla Guardians API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: artisanId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ARTISANS
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # DELIVERY TRACKING
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        productCount:
          type: integer
      required: [id, name, slug]

    # PRODUCT
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        culturalSignificance:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        stock:
          type: integer
        sku:
          type: ["string", "null"]
        categoryId:
          type: integer
        artisanId:
          type: integer
        images:
          type: array
          items:
            type: string
        materials:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        dimensions:
          type: ["string", "null"]
        featured:
          type: boolean
        active:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        artisan:
          $ref: "#/components/schemas/Artisan"
        category:
          $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, description, price, stock, categoryId, artisanId, images, featured, active, reviewCount, createdAt]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, totalPages]
    ProductInput:
      type: object
      required: [name, description, price, stock, categoryId, artisanId, images]
      properties:
        name:
          type: string
        description:
          type: string
        culturalSignificance:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        sku:
          type: string
        categoryId:
          type: integer
        artisanId:
          type: integer
        images:
          type: array
          items:
            type: string
        materials:
          type: string
        weight:
          type: number
        dimensions:
          type: string
        featured:
          type: boolean
        active:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        culturalSignificance:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        materials:
          type: string
        featured:
          type: boolean
        active:
          type: boolean

    # ARTISAN
    Artisan:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        photo:
          type: ["string", "null"]
        biography:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        story:
          type: ["string", "null"]
        galleryImages:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        productCount:
          type: integer
        totalSales:
          type: integer
        averageRating:
          type: ["number", "null"]
        totalDonations:
          type: number
        isConservationAmbassador:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, skills, galleryImages, productCount, totalSales, totalDonations, isConservationAmbassador, createdAt]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, packed, shipped, out_for_delivery, delivered, cancelled, refunded]
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        shippingAddress:
          type: string
        shippingType:
          type: string
          enum: [global, pickup]
        shippingCarrier:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, items, subtotal, shippingCost, total, status, paymentMethod, paymentStatus, shippingAddress, shippingType, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required: [id, productId, productName, productImage, quantity, price, subtotal]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, totalPages]
    OrderInput:
      type: object
      required: [items, shippingAddress, shippingType, paymentMethod]
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required: [productId, quantity]
        shippingAddress:
          type: string
        shippingType:
          type: string
          enum: [global, pickup]
        paymentMethod:
          type: string
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
        paymentStatus:
          type: string
        trackingNumber:
          type: string
        shippingCarrier:
          type: string
        notes:
          type: string

    # DELIVERY TRACKING
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, message, isRead, createdAt]

    # CONVERSATION / MESSAGE
...