# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gorilla Guardians API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  culturalSignificance?: string;
  price: number;
  discountPrice?: number;
  stock: number;
  sku?: string;
  categoryId: number;
  artisanId: number;
  images: string[];
  materials?: string;
  weight?: number;
  dimensions?: string;
  featured?: boolean;
  active?: boolean;
}